# backend/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: This is the backend API for the Time to Dry project.
  title: Time To Dry API
  version: "1.0"
paths:
...
  /api/ttd/status/check:
    get:
      description: Returns whether a given test_id is still collecting data or completed.
      parameters:
      - description: Test ID to check
        in: query
        name: test_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing or invalid test_id
          schema:
            type: string
        "404":
          description: No records found
          schema:
            type: string
      summary: Check specific test status
      tags:
      - Test
swagger: "2.0"